feat: implement buildCoseSign1 function for creating COSE_Sign1 structure
