+++ src/utils/cardano.utils.ts
@@ -413,3 +413,49 @@ export const submitTransaction = async (
     );
   }
 };
+
+/**
+ * Builds a CIP-30 compliant COSE_Sign1 structure from a Fireblocks signature
+ * @param message - The original message that was signed (UTF-8 string)
+ * @param fullSig - The full signature from Fireblocks (hex string)
+ * @param publicKey - The public key from Fireblocks (hex string)
+ * @returns Object containing the COSE_Sign1 hex and public key
+ */
+export const buildCoseSign1 = async (
+  message: string,
+  fullSig: string,
+): Promise<string> => {
+  try {
+    const { MSL } = await import("cardano-web3-js");
+
+    // Create protected headers with EdDSA algorithm
+    const protectedHeaders = MSL.HeaderMap.new();
+    protectedHeaders.set_algorithm_id(
+      MSL.Label.from_algorithm_id(MSL.AlgorithmId.EdDSA)
+    );
+
+    // Build headers structure
+    const protectedSerialized = MSL.ProtectedHeaderMap.new(protectedHeaders);
+    const unprotectedHeaders = MSL.HeaderMap.new();
+    const headers = MSL.Headers.new(protectedSerialized, unprotectedHeaders);
+
+    // Convert message to bytes
+    const messageBytes = new Uint8Array(Buffer.from(message, "utf8"));
+
+    // Build COSE_Sign1 structure
+    const builder = MSL.COSESign1Builder.new(headers, messageBytes, false);
+    const signatureBytes = new Uint8Array(Buffer.from(fullSig, "hex"));
+    const coseSign1 = builder.build(signatureBytes);
+
+    // Convert to hex
+    const coseSign1Hex = Buffer.from(coseSign1.to_bytes()).toString("hex");
+
+    return coseSign1Hex;
+  } catch (error) {
+    throw new Error(
+      `Failed to build COSE_Sign1: ${
+        error instanceof Error ? error.message : error
+      }`
+    );
+  }
+};
